requirements.txt
flask
flask-cors
transformers
python-dotenv
torch
httpx
slack_sdk
